Calculate Discount Weights

Computes discount weights for a sequence of cycles given a discount rate expressed per annum and the length of each cycle. These weights can be multiplied by costs or QALYs to obtain discounted values.

Usage

calc_discount_wts(discount_rate, num_cycles, cycle_length)

Arguments

discount_rate

Annual discount rate expressed as a fraction (e.g. 0.03).

num_cycles

Integer giving the number of cycles in the simulation.

cycle_length

Length of each cycle in years.

Value

A numeric vector of discount weights of length num_cycles + 1.

Examples

calc_discount_wts(
  discount_rate = 0.03,
  num_cycles = 5, 
  cycle_length = 1
)
#> [1] 1.0000000 0.9708738 0.9425959 0.9151417 0.8884870 0.8626088